refactor: move all core SOR algo code to router dir
