+++ src/index.ts
@@ -5,9 +5,7 @@ export { getCostOutputToken } from './costToken';
 export { SOR } from './wrapper';
 export * from './config';
 export * from './types';
-export * from './helpersClass';
 export * from './pools';
-export * from './sorClass';
 export * from './frontendHelpers/weightedHelpers';
 export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
 export { scale, bnum } from './utils/bignumber';
