+++ src/router/helpersClass.ts
@@ -1,4 +1,4 @@
-import { BigNumber } from './utils/bignumber';
+import { BigNumber } from '../utils/bignumber';
 import {
     NewPath,
     PoolDictionary,
@@ -7,15 +7,9 @@ import {
     PoolBase,
     PoolPairBase,
     PoolTypes,
-    SwapV2,
-    Swap,
-    SwapInfo,
-} from './types';
-import { bnum, scale, ZERO, INFINITY } from './utils/bignumber';
-import { INFINITESIMAL } from './config';
-import { ZERO_ADDRESS, Lido, getStEthRate } from './index';
-import { WETHADDR } from './constants';
-import { BaseProvider } from '@ethersproject/providers';
+} from '../types';
+import { ZERO, INFINITY } from '../utils/bignumber';
+import { INFINITESIMAL } from '../config';
 
 export function getHighestLimitAmountsForPaths(
     paths: NewPath[],
@@ -510,343 +504,4 @@ export function EVMgetOutputAmountSwap(
     pool.updateTokenBalanceForPool(tokenOut, balanceOut.minus(amount));
 
     return returnAmount;
-}
-
-export function formatSwaps(
-    swapsOriginal: Swap[][],
-    swapType: SwapTypes,
-    swapAmount: BigNumber,
-    tokenIn: string,
-    tokenOut: string,
-    returnAmount: BigNumber,
-    returnAmountConsideringFees: BigNumber,
-    marketSp: BigNumber
-): SwapInfo {
-    const tokenAddressesSet: Set<string> = new Set();
-
-    const swaps: Swap[][] = JSON.parse(JSON.stringify(swapsOriginal));
-
-    let tokenInDecimals: number;
-    let tokenOutDecimals: number;
-
-    const swapInfo: SwapInfo = {
-        tokenAddresses: [],
-        swaps: [],
-        swapAmount: ZERO,
-        swapAmountForSwaps: ZERO,
-        returnAmount: ZERO,
-        returnAmountConsideringFees: ZERO,
-        returnAmountFromSwaps: ZERO,
-        tokenIn: '',
-        tokenOut: '',
-        marketSp: marketSp,
-    };
-
-    if (swaps.length === 0) {
-        return swapInfo;
-    }
-
-    swaps.forEach(sequence => {
-        sequence.forEach(swap => {
-            if (swap.tokenIn === tokenIn)
-                tokenInDecimals = swap.tokenInDecimals;
-
-            if (swap.tokenOut === tokenOut)
-                tokenOutDecimals = swap.tokenOutDecimals;
-
-            tokenAddressesSet.add(swap.tokenIn);
-            tokenAddressesSet.add(swap.tokenOut);
-        });
-    });
-
-    const tokenArray = [...tokenAddressesSet];
-
-    if (swapType === SwapTypes.SwapExactIn) {
-        const swapsV2: SwapV2[] = [];
-
-        let totalSwapAmount = ZERO;
-        /*
-         * Multihop swaps can be executed by passing an`amountIn` value of zero for a swap.This will cause the amount out
-         * of the previous swap to be used as the amount in of the current one.In such a scenario, `tokenIn` must equal the
-         * previous swap's `tokenOut`.
-         * */
-        swaps.forEach(sequence => {
-            sequence.forEach((swap, i) => {
-                let amountScaled = '0'; // amount will be 0 for second swap in multihop swap
-                if (i == 0) {
-                    // First swap so should have a value for both single and multihop
-                    amountScaled = scale(
-                        bnum(swap.swapAmount),
-                        swap.tokenInDecimals
-                    )
-                        .decimalPlaces(0, 1)
-                        .toString();
-                    totalSwapAmount = totalSwapAmount.plus(amountScaled);
-                }
-
-                const inIndex = tokenArray.indexOf(swap.tokenIn);
-                const outIndex = tokenArray.indexOf(swap.tokenOut);
-                const swapV2: SwapV2 = {
-                    poolId: swap.pool,
-                    assetInIndex: inIndex,
-                    assetOutIndex: outIndex,
-                    amount: amountScaled,
-                    userData: '0x',
-                };
-
-                swapsV2.push(swapV2);
-            });
-        });
-
-        // We need to account for any rounding losses by adding dust to first path
-        const swapAmountScaled = scale(swapAmount, tokenInDecimals);
-        const dust = swapAmountScaled.minus(totalSwapAmount).dp(0, 0);
-        if (dust.gt(0))
-            swapsV2[0].amount = bnum(swapsV2[0].amount)
-                .plus(dust)
-                .toString();
-
-        swapInfo.swapAmount = swapAmountScaled;
-        // Using this split to remove any decimals
-        swapInfo.returnAmount = bnum(
-            scale(returnAmount, tokenOutDecimals)
-                .toString()
-                .split('.')[0]
-        );
-        swapInfo.returnAmountConsideringFees = bnum(
-            scale(returnAmountConsideringFees, tokenOutDecimals)
-                .toString()
-                .split('.')[0]
-        );
-        swapInfo.swaps = swapsV2;
-    } else {
-        let swapsV2: SwapV2[] = [];
-        let totalSwapAmount = ZERO;
-        /*
-        SwapExactOut will have order reversed in V2.
-        v1 = [[x, y]], [[a, b]]
-        v2 = [y, x, b, a]
-        */
-        swaps.forEach(sequence => {
-            if (sequence.length > 2)
-                throw new Error(
-                    'Multihop with more than 2 swaps not supported'
-                );
-
-            const sequenceSwaps = [];
-            sequence.forEach((swap, i) => {
-                const inIndex = tokenArray.indexOf(swap.tokenIn);
-                const outIndex = tokenArray.indexOf(swap.tokenOut);
-                const swapV2: SwapV2 = {
-                    poolId: swap.pool,
-                    assetInIndex: inIndex,
-                    assetOutIndex: outIndex,
-                    amount: '0', // For a multihop the first swap in sequence should be last in order and have amt = 0
-                    userData: '0x',
-                };
-
-                if (i == 0 && sequence.length > 1) {
-                    sequenceSwaps[1] = swapV2; // Make the swap the last in V2 order for the sequence
-                } else {
-                    const amountScaled = scale(
-                        bnum(swap.swapAmount),
-                        swap.tokenOutDecimals
-                    )
-                        .decimalPlaces(0, 1)
-                        .toString();
-                    totalSwapAmount = totalSwapAmount.plus(amountScaled);
-                    swapV2.amount = amountScaled; // Make the swap the first in V2 order for the sequence with the value
-                    sequenceSwaps[0] = swapV2;
-                }
-            });
-
-            swapsV2 = swapsV2.concat(sequenceSwaps);
-        });
-
-        // We need to account for any rounding losses by adding dust to first path
-        const swapAmountScaled = scale(swapAmount, tokenOutDecimals);
-        const dust = swapAmountScaled.minus(totalSwapAmount).dp(0, 0);
-        if (dust.gt(0))
-            swapsV2[0].amount = bnum(swapsV2[0].amount)
-                .plus(dust)
-                .toString();
-
-        swapInfo.swapAmount = swapAmountScaled;
-        // Using this split to remove any decimals
-        swapInfo.returnAmount = bnum(
-            scale(returnAmount, tokenInDecimals)
-                .toString()
-                .split('.')[0]
-        );
-        swapInfo.returnAmountConsideringFees = bnum(
-            scale(returnAmountConsideringFees, tokenInDecimals)
-                .toString()
-                .split('.')[0]
-        );
-        swapInfo.swaps = swapsV2;
-    }
-
-    swapInfo.tokenAddresses = tokenArray;
-    swapInfo.tokenIn = tokenIn;
-    swapInfo.tokenOut = tokenOut;
-    return swapInfo;
-}
-
-export interface WrappedInfo {
-    swapAmountOriginal: BigNumber;
-    swapAmountForSwaps: BigNumber;
-    tokenIn: TokenInfo;
-    tokenOut: TokenInfo;
-}
-
-export interface TokenInfo {
-    addressOriginal: string;
-    addressForSwaps: string;
-    wrapType: WrapTypes;
-    rate: BigNumber;
-}
-
-export enum WrapTypes {
-    None,
-    ETH,
-    stETH,
-}
-
-export async function getWrappedInfo(
-    provider: BaseProvider,
-    swapType: SwapTypes,
-    tokenIn: string,
-    tokenOut: string,
-    chainId: number,
-    swapAmount: BigNumber
-): Promise<WrappedInfo> {
-    // The Subgraph returns tokens in lower case format so we must match this
-    tokenIn = tokenIn.toLowerCase();
-    tokenOut = tokenOut.toLowerCase();
-
-    let swapAmountForSwaps = swapAmount;
-
-    let tokenInForSwaps = tokenIn;
-    let tokenInWrapType = WrapTypes.None;
-    let tokenOutForSwaps = tokenOut;
-    let tokenOutWrapType = WrapTypes.None;
-    let tokenInRate = bnum(1);
-    let tokenOutRate = bnum(1);
-
-    // Handle ETH wrapping
-    if (tokenIn === ZERO_ADDRESS) {
-        tokenInForSwaps = WETHADDR[chainId].toLowerCase();
-        tokenInWrapType = WrapTypes.ETH;
-    }
-    if (tokenOut === ZERO_ADDRESS) {
-        tokenOutForSwaps = WETHADDR[chainId].toLowerCase();
-        tokenOutWrapType = WrapTypes.ETH;
-    }
-
-    // Handle stETH wrapping
-    if (tokenIn === Lido.stETH[chainId]) {
-        tokenInForSwaps = Lido.wstETH[chainId];
-        tokenInWrapType = WrapTypes.stETH;
-        const rate = await getStEthRate(provider, chainId);
-        tokenInRate = rate;
-        if (swapType === SwapTypes.SwapExactIn)
-            swapAmountForSwaps = swapAmount.times(rate).dp(18);
-    }
-    if (tokenOut === Lido.stETH[chainId]) {
-        tokenOutForSwaps = Lido.wstETH[chainId];
-        tokenOutWrapType = WrapTypes.stETH;
-        const rate = await getStEthRate(provider, chainId);
-        tokenOutRate = rate;
-        if (swapType === SwapTypes.SwapExactOut)
-            swapAmountForSwaps = swapAmount.times(rate).dp(18);
-    }
-
-    return {
-        swapAmountOriginal: swapAmount,
-        swapAmountForSwaps: swapAmountForSwaps,
-        tokenIn: {
-            addressOriginal: tokenIn,
-            addressForSwaps: tokenInForSwaps,
-            wrapType: tokenInWrapType,
-            rate: tokenInRate,
-        },
-        tokenOut: {
-            addressOriginal: tokenOut,
-            addressForSwaps: tokenOutForSwaps,
-            wrapType: tokenOutWrapType,
-            rate: tokenOutRate,
-        },
-    };
-}
-
-export function setWrappedInfo(
-    swapInfo: SwapInfo,
-    swapType: SwapTypes,
-    wrappedInfo: WrappedInfo,
-    chainId: number
-): SwapInfo {
-    if (swapInfo.swaps.length === 0) return swapInfo;
-
-    swapInfo.tokenIn = wrappedInfo.tokenIn.addressOriginal;
-    swapInfo.tokenOut = wrappedInfo.tokenOut.addressOriginal;
-
-    // replace weth with ZERO/ETH in assets for Vault to handle ETH directly
-    if (
-        wrappedInfo.tokenIn.wrapType === WrapTypes.ETH ||
-        wrappedInfo.tokenOut.wrapType === WrapTypes.ETH
-    ) {
-        let wethIndex = -1;
-        swapInfo.tokenAddresses.forEach((addr, i) => {
-            if (addr.toLowerCase() === WETHADDR[chainId].toLowerCase())
-                wethIndex = i;
-        });
-        if (wethIndex !== -1) swapInfo.tokenAddresses[wethIndex] = ZERO_ADDRESS;
-    }
-
-    // Handle stETH swap amount scaling
-    if (
-        (wrappedInfo.tokenIn.wrapType === WrapTypes.stETH &&
-            swapType === SwapTypes.SwapExactIn) ||
-        (wrappedInfo.tokenOut.wrapType === WrapTypes.stETH &&
-            swapType === SwapTypes.SwapExactOut)
-    ) {
-        swapInfo.swapAmountForSwaps = scale(
-            wrappedInfo.swapAmountForSwaps,
-            18
-        ).dp(0); // Always 18 because wstETH
-        swapInfo.swapAmount = scale(wrappedInfo.swapAmountOriginal, 18).dp(0);
-    } else {
-        // Should be same when standard tokens and swapAmount should already be scaled
-        swapInfo.swapAmountForSwaps = swapInfo.swapAmount;
-    }
-
-    // Return amount from swaps will only be different if token has an exchangeRate
-    swapInfo.returnAmountFromSwaps = swapInfo.returnAmount;
-
-    // SwapExactIn, stETH out, returnAmount is stETH amount out, returnAmountForSwaps is wstETH amount out
-    if (
-        swapType === SwapTypes.SwapExactIn &&
-        wrappedInfo.tokenOut.wrapType === WrapTypes.stETH
-    ) {
-        swapInfo.returnAmount = swapInfo.returnAmount
-            .div(wrappedInfo.tokenOut.rate)
-            .dp(0);
-        swapInfo.returnAmountConsideringFees = swapInfo.returnAmountConsideringFees
-            .div(wrappedInfo.tokenOut.rate)
-            .dp(0);
-    }
-
-    // SwapExactOut, stETH in, returnAmount us stETH amount in, returnAmountForSwaps is wstETH amount in
-    if (
-        swapType === SwapTypes.SwapExactOut &&
-        wrappedInfo.tokenIn.wrapType === WrapTypes.stETH
-    ) {
-        swapInfo.returnAmount = swapInfo.returnAmount
-            .div(wrappedInfo.tokenIn.rate)
-            .dp(0);
-        swapInfo.returnAmountConsideringFees = swapInfo.returnAmountConsideringFees
-            .div(wrappedInfo.tokenIn.rate)
-            .dp(0);
-    }
-    return swapInfo;
 }
